+++ NEWS.md
@@ -2,6 +2,7 @@
 - Fix `make -j` with GNU make (#440).
 - Subtraction from zero could return the wrong result in some
   circumstances.
+- Fix incorrect code generation for access-to-array.
 
 ## Version 1.6.1 - 2022-02-05
 - Fix compatibility with BSD make (#440).